Fix factory automation not being disabled when it should
I think it would be way more concise if an additional flag is added to represent each of 3 planetary options for automation.
I also think that using autofab_automation_off flag to represent a state where planet factory automation is turned on is not very good. Maybe it's a bug, not entirely sure.
Regardless, my goal was to just fix a bug without introducing any other side effects, so this commit should do for now.